requirements.txt
tqdm
torch
numpy
pillow